+++ annis-gui/src/main/java/annis/gui/objects/PagedResultQuery.java
@@ -31,7 +31,7 @@ public class PagedResultQuery extends ContextualizedQuery implements Cloneable
 {
   private final static Logger log = LoggerFactory.getLogger(PagedResultQuery.class);
   
-  private int offset;
+  private long offset;
   private int limit;
   private OrderType order = OrderType.ascending;
 
@@ -40,7 +40,7 @@ public class PagedResultQuery extends ContextualizedQuery implements Cloneable
     
   }
 
-  public PagedResultQuery(int contextLeft, int contextRight, int offset,
+  public PagedResultQuery(int contextLeft, int contextRight, long offset,
     int limit, String segmentation, String query,
     Set<String> corpora)
   {
@@ -54,12 +54,12 @@ public class PagedResultQuery extends ContextualizedQuery implements Cloneable
     this.limit = limit;
   }
 
-  public int getOffset()
+  public long getOffset()
   {
     return offset;
   }
 
-  public void setOffset(int offset)
+  public void setOffset(long offset)
   {
     this.offset = offset;
   }
